Move code to share Log interfaces and lint it
